requirements.txt
torch 
torchvision 
torchaudio
torchsummary
numpy
fonttools
tqdm
matplotlib
imageio
Pillow
gradio
opencv-python

# pip install -r requirements.txt